Growth boost a = 0.025

+++ run_test/execute.py
@@ -222,6 +222,10 @@ if __name__=="__main__":
     out_dir = get_sample_dir(folder_name)
     os.makedirs(out_dir, exist_ok=True)
     fiS = f"{out_dir}/{out_name}"
+    if os.path.isfile(fiS):
+        raise SystemExit(
+            f'File {fiS} already existed. Please delete or recheck'
+        )
     log.info(f'Saving at {fiS}')
     
     with open(fiS, 'wb') as file:
